updated strings/parse-position

+++ exercises/04-strings/solutions.py
@@ -41,14 +41,14 @@ def parse_position_x(string):
     without_parentheses = string[1:-1]
     index = without_parentheses.find(',')
     left = without_parentheses[:index]
-    return int(left)
+    return float(left)
 
 
 def parse_position_y(string):
     without_parentheses = string[1:-1]
     index = without_parentheses.find(',')
     right = without_parentheses[index+2:]
-    return int(right)
+    return float(right)
 
 
 def is_capitalized(string):
